requirements.txt
alembic==1.10.4
black==23.3.0
blinker==1.6.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
filelock==3.13.1
flake8==6.0.0
Flask==2.3.2
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
flask-paginate==2023.10.24
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fsspec==2023.12.2
ftfy==6.1.3
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mpmath==1.3.0
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
open-clip-torch==2.24.0
packaging==23.1
pathspec==0.11.1
Pillow==10.1.0
platformdirs==3.5.1
protobuf==4.25.2
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
reportlab==4.0.7
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.2
sentencepiece==0.1.99
sniffio==1.3.0
SQLAlchemy==2.0.13
sympy==1.12
timm==0.9.12
tokenizers==0.15.1
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
ultralytics==8.1.28
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==2.3.4
WTForms==3.0.1
PyQt5==5.15.10
flet==0.22.0
opencv-python==4.8.1.78
